requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
dnspython>=2.7.0
django-redis>=5.2.1